fix: cast rounded money value to int
Ensure the rounded money value is properly cast to an integer.
This resolves potential issues with data type consistency.

+++ app/Casts/MoneyCast.php
@@ -24,6 +24,6 @@ class MoneyCast implements CastsAttributes
      */
     public function set(Model $model, string $key, mixed $value, array $attributes): int
     {
-        return round(floatval($value) * 100);
+        return (int) round(floatval($value) * 100);
     }
 }
